% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{JCF_palette}
\alias{JCF_palette}
\title{Color Palette}
\usage{
JCF_palette(colors)
}
\arguments{
\item{colors}{Numeric. The number of colors to return.}
}
\description{
Use this function to get a vector of colors.
}
