% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{palette_minimal}
\alias{palette_minimal}
\title{Color Palette}
\usage{
palette_minimal(colors)
}
\arguments{
\item{colors}{Numeric. The number of colors to return.}
}
\description{
Use this function to get a vector of colors.
}
