% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_minimal}
\alias{theme_minimal}
\title{Plot Theme}
\usage{
theme_minimal(size = 12)
}
\arguments{
\item{size}{Numeric. The size of the font to use.}
}
\description{
Use this function to add the theme to a plot.
}
