% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{titles_minimal}
\alias{titles_minimal}
\title{Plot Titles}
\usage{
titles_minimal(x = NULL, y = NULL, title = NULL)
}
\arguments{
\item{x}{String. The title for the x-axis.}

\item{y}{String. The title for the y-axis.}

\item{main}{String. The main title for the plot.}
}
\description{
Use this function to add axis titles and a plot title.
}
